requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
pymupdf
azure-ai-documentintelligence
azure-identity
azure-storage-blob
python-docx
python-pptx
jinja2
pytest
pytest-cov
pytest-asyncio
